+++ app/src/main/java/org/ole/planet/myplanet/lite/SurveyWizardFragment.kt
@@ -80,50 +80,50 @@ import androidx.core.graphics.toColorInt
 
 class SurveyWizardFragment : Fragment(R.layout.fragment_survey_wizard) {
 
-    private var document: SurveyDocument? = null
-    private var questions: List<SurveyQuestion> = emptyList()
-    private var teamId: String? = null
-    private var teamName: String? = null
-    private var courseId: String? = null
-    private var baseUrlOverride: String? = null
-    private var includeUserContext: Boolean = true
-    private var steps: List<WizardStep> = emptyList()
-    private var currentIndex = 0
-    private val answers: MutableMap<Int, SurveyAnswer> = mutableMapOf()
-    private var activeCollector: (() -> Boolean)? = null
-    private val respondent = SurveyRespondent()
-    private var birthDateSelection: Long? = null
-    private var includeOptionalDetails = false
-    private val submissionRepository = DashboardSurveySubmissionsRepository()
-    private var baseUrl: String? = null
-    private var credentials: StoredCredentials? = null
-    private var sessionCookie: String? = null
-    private var serverCode: String? = null
-    private var parentCode: String? = null
-    private var isExam: Boolean = false
-    private var startTimeMillis: Long = System.currentTimeMillis()
-    private var isSubmitting = false
-    private lateinit var translationManager: SurveyTranslationManager
-    private var questionTranslations: Map<Int, TranslatedQuestion> = emptyMap()
-    private var targetSurveyLanguage: String? = null
-    private var detectedSurveyLanguage: String? = null
-    private var translatedTitle: String? = null
-    private var translatedDescription: String? = null
-    private val localSurveyRepository by lazy { DashboardLocalSurveyRepository(requireContext()) }
+    internal var document: SurveyDocument? = null
+    internal var questions: List<SurveyQuestion> = emptyList()
+    internal var teamId: String? = null
+    internal var teamName: String? = null
+    internal var courseId: String? = null
+    internal var baseUrlOverride: String? = null
+    internal var includeUserContext: Boolean = true
+    internal var steps: List<WizardStep> = emptyList()
+    internal var currentIndex = 0
+    internal val answers: MutableMap<Int, SurveyAnswer> = mutableMapOf()
+    internal var activeCollector: (() -> Boolean)? = null
+    internal val respondent = SurveyRespondent()
+    internal var birthDateSelection: Long? = null
+    internal var includeOptionalDetails = false
+    internal val submissionRepository = DashboardSurveySubmissionsRepository()
+    internal var baseUrl: String? = null
+    internal var credentials: StoredCredentials? = null
+    internal var sessionCookie: String? = null
+    internal var serverCode: String? = null
+    internal var parentCode: String? = null
+    internal var isExam: Boolean = false
+    internal var startTimeMillis: Long = System.currentTimeMillis()
+    internal var isSubmitting = false
+    internal lateinit var translationManager: SurveyTranslationManager
+    internal var questionTranslations: Map<Int, TranslatedQuestion> = emptyMap()
+    internal var targetSurveyLanguage: String? = null
+    internal var detectedSurveyLanguage: String? = null
+    internal var translatedTitle: String? = null
+    internal var translatedDescription: String? = null
+    internal val localSurveyRepository by lazy { DashboardLocalSurveyRepository(requireContext()) }
 
-    private lateinit var titleView: TextView
-    private lateinit var descriptionView: TextView
-    private lateinit var counterView: TextView
-    private lateinit var progressBar: ProgressBar
-    private lateinit var translationOverlay: View
-    private lateinit var translationProgressBar: ProgressBar
-    private lateinit var translationNoticeView: TextView
-    private var translationApplied: Boolean = false
-    private lateinit var questionBodyView: TextView
-    private lateinit var questionScrollView: ScrollView
-    private lateinit var questionContainer: LinearLayout
-    private lateinit var previousButton: Button
-    private lateinit var nextButton: Button
+    internal lateinit var titleView: TextView
+    internal lateinit var descriptionView: TextView
+    internal lateinit var counterView: TextView
+    internal lateinit var progressBar: ProgressBar
+    internal lateinit var translationOverlay: View
+    internal lateinit var translationProgressBar: ProgressBar
+    internal lateinit var translationNoticeView: TextView
+    internal var translationApplied: Boolean = false
+    internal lateinit var questionBodyView: TextView
+    internal lateinit var questionScrollView: ScrollView
+    internal lateinit var questionContainer: LinearLayout
+    internal lateinit var previousButton: Button
+    internal lateinit var nextButton: Button
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -180,64 +180,6 @@ class SurveyWizardFragment : Fragment(R.layout.fragment_survey_wizard) {
         setupNavigationButtons()
     }
 
-    private fun setupNavigationButtons() {
-        previousButton.setOnClickListener {
-            activeCollector?.invoke()
-            if (currentIndex > 0) {
-                currentIndex -= 1
-                showStep(currentIndex)
-            }
-        }
-
-        nextButton.setOnClickListener {
-            val collector = activeCollector ?: return@setOnClickListener
-            if (collector()) {
-                if (currentIndex < steps.lastIndex) {
-                    currentIndex += 1
-                    showStep(currentIndex)
-                } else {
-                    submitSurvey()
-                }
-            }
-        }
-    }
-
-    private fun setupInsets(view: View) {
-        val contentView: View = view.findViewById(R.id.surveyWizardContent)
-        val initialPaddingStart = contentView.paddingStart
-        val initialPaddingTop = contentView.paddingTop
-        val initialPaddingEnd = contentView.paddingEnd
-        val initialPaddingBottom = contentView.paddingBottom
-
-        ViewCompat.setOnApplyWindowInsetsListener(contentView) { content, insets ->
-            val systemInsets = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            val imeInsets = insets.getInsets(WindowInsetsCompat.Type.ime())
-            val bottomInset = maxOf(systemInsets.bottom, imeInsets.bottom)
-            content.setPadding(
-                initialPaddingStart,
-                initialPaddingTop,
-                initialPaddingEnd,
-                initialPaddingBottom + bottomInset
-            )
-            insets
-        }
-    }
-
-    private fun bindViews(view: View) {
-        titleView = view.findViewById(R.id.surveyWizardTitle)
-        descriptionView = view.findViewById(R.id.surveyWizardDescription)
-        counterView = view.findViewById(R.id.surveyWizardCounter)
-        progressBar = view.findViewById(R.id.surveyWizardProgress)
-        translationOverlay = view.findViewById(R.id.surveyWizardTranslationOverlay)
-        translationProgressBar = view.findViewById(R.id.surveyWizardTranslationProgress)
-        translationNoticeView = view.findViewById(R.id.surveyWizardTranslationNotice)
-        questionBodyView = view.findViewById(R.id.surveyWizardQuestionBody)
-        questionScrollView = view.findViewById(R.id.surveyWizardQuestionScrollView)
-        questionContainer = view.findViewById(R.id.surveyWizardQuestionContainer)
-        previousButton = view.findViewById(R.id.surveyWizardPreviousButton)
-        nextButton = view.findViewById(R.id.surveyWizardNextButton)
-    }
-
     override fun onStart() {
         super.onStart()
         viewLifecycleOwner.lifecycleScope.launch {
@@ -245,1630 +187,6 @@ class SurveyWizardFragment : Fragment(R.layout.fragment_survey_wizard) {
                 flushPendingSurveySubmissions()
             }
         }
-    }
-
-    private suspend fun initializeSession() {
-        val context = requireContext().applicationContext
-        baseUrl = baseUrlOverride ?: DashboardServerPreferences.getServerBaseUrl(context)
-        credentials = if (includeUserContext) {
-            ProfileCredentialsStore.getStoredCredentials(context)
-        } else {
-            null
-        }
-        serverCode = if (includeUserContext) DashboardServerPreferences.getServerCode(context) else null
-        parentCode = if (includeUserContext) DashboardServerPreferences.getServerParentCode(context) else null
-        baseUrl?.takeIf { includeUserContext }?.let { base ->
-            val authService = AuthDependencies.provideAuthService(context, base)
-            sessionCookie = authService.getStoredToken()
-        }
-    }
-
-    private suspend fun attemptSurveyTranslation() {
-        val survey = document ?: return
-        val translationPreferenceEnabled = DashboardActivity.isSurveyTranslationEnabled(requireContext())
-        val translationConsentAccepted = DashboardActivity.isSurveyTranslationConsentAccepted(requireContext())
-        val translationAllowed = translationPreferenceEnabled && translationConsentAccepted
-        if (!translationAllowed) {
-            translatedTitle = null
-            translatedDescription = null
-            questionTranslations = emptyMap()
-            translationApplied = false
-            updateTranslationNotice(
-                showConsentNotice = translationPreferenceEnabled,
-                showAppliedNotice = false,
-            )
-            setTranslationInProgress(false)
-            return
-        }
-        val base = baseUrl?.takeIf { it.isNotBlank() }
-            ?: DashboardServerPreferences.getServerBaseUrl(requireContext().applicationContext)
-            ?: return
-        val appLocales = AppCompatDelegate.getApplicationLocales()
-        val targetLanguage = appLocales[0]?.language
-            ?.takeIf { it.isNotBlank() }
-            ?: Locale.getDefault().language
-        targetSurveyLanguage = targetLanguage
-        val detectedLanguage = translationManager.detectSurveyLanguage(survey)
-        detectedSurveyLanguage = detectedLanguage
-        val shouldShowTranslationProgress = detectedLanguage != null &&
-            !detectedLanguage.equals(targetLanguage, ignoreCase = true)
-        setTranslationInProgress(shouldShowTranslationProgress)
-        val result = translationManager.translateSurvey(
-            base,
-            survey,
-            targetLanguage,
-            detectedLanguage,
-        )
-        setTranslationInProgress(false)
-        translatedTitle = result.titleTranslation?.takeIf { it.isNotBlank() }
-        translatedDescription = result.descriptionTranslation?.takeIf { it.isNotBlank() }
-        applySurveyTranslations()
-        val translations = result.translations.filterValues { it.hasTranslation }
-        if (translations.isNotEmpty()) {
-            questionTranslations = translations
-            if (isAdded) {
-                val currentStepValue = steps.getOrNull(currentIndex)
-                if (currentStepValue is WizardStep.Question) {
-                    showStep(currentIndex)
-                }
-            }
-        }
-        translationApplied = translatedTitle != null || translatedDescription != null ||
-            translations.isNotEmpty() || (detectedSurveyLanguage != null &&
-            targetSurveyLanguage != null &&
-            !detectedSurveyLanguage.equals(targetSurveyLanguage, ignoreCase = true))
-        updateTranslationNotice(
-            showConsentNotice = false,
-            showAppliedNotice = translationApplied,
-        )
-    }
-
-    private fun applySurveyTranslations() {
-        val survey = document ?: return
-        val titleText = translatedTitle?.takeIf { it.isNotBlank() }
-            ?: survey.name.orEmpty()
-        titleView.text = titleText
-        val descriptionText = translatedDescription?.takeIf { it.isNotBlank() }
-            ?: survey.description.orEmpty()
-        descriptionView.text = descriptionText
-        updateDescriptionVisibility(currentIndex)
-    }
-
-    private fun updateTranslationNotice(showConsentNotice: Boolean, showAppliedNotice: Boolean) {
-        when {
-            showConsentNotice -> {
-                translationNoticeView.text = getString(R.string.dashboard_survey_translation_notice)
-                translationNoticeView.isVisible = true
-            }
-            showAppliedNotice -> {
-                translationNoticeView.text = getString(R.string.dashboard_survey_translation_applied_notice)
-                translationNoticeView.isVisible = true
-            }
-            else -> {
-                translationNoticeView.isVisible = false
-            }
-        }
-    }
-
-    private fun setTranslationInProgress(inProgress: Boolean) {
-        translationOverlay.isVisible = inProgress
-        translationProgressBar.isVisible = inProgress
-        if (inProgress) {
-            translationOverlay.bringToFront()
-        }
-        updateNavigationEnabled()
-    }
-
-    private fun updateNavigationEnabled() {
-        val enabled = !isSubmitting && !translationOverlay.isVisible
-        previousButton.isEnabled = enabled && currentIndex > 0
-        nextButton.isEnabled = enabled
-    }
-
-    private fun showStep(index: Int) {
-        val step = steps[index]
-        progressBar.max = steps.size
-        counterView.text = getString(R.string.dashboard_survey_wizard_step_counter, index + 1, steps.size)
-        updateDescriptionVisibility(index)
-        questionBodyView.text = when (step) {
-            WizardStep.Basics -> getString(R.string.dashboard_survey_wizard_participant_basics_title)
-            WizardStep.Names -> getString(R.string.dashboard_survey_wizard_names_title)
-            WizardStep.BirthDate -> getString(R.string.dashboard_survey_wizard_birthdate_title)
-            WizardStep.Contact -> getString(R.string.dashboard_survey_wizard_contact_title)
-            WizardStep.LanguageLevel -> getString(R.string.dashboard_survey_wizard_language_level_title)
-            is WizardStep.Question -> questionTranslations[step.questionIndex]?.body
-                ?: step.question.body.orEmpty()
-        }
-        questionContainer.removeAllViews()
-        val (renderedView, collector) = renderStep(step)
-        questionContainer.addView(renderedView)
-        activeCollector = collector
-        progressBar.progress = index + 1
-        val isLast = index == steps.lastIndex
-        nextButton.text = if (isLast) {
-            getString(R.string.dashboard_survey_wizard_finish)
-        } else {
-            getString(R.string.dashboard_survey_wizard_next)
-        }
-        updateNavigationEnabled()
-    }
-
-    private fun updateDescriptionVisibility(stepIndex: Int) {
-        descriptionView.isVisible = stepIndex == 0 && descriptionView.text.isNotBlank()
-    }
-
-    private fun renderStep(step: WizardStep): Pair<View, () -> Boolean> {
-        return when (step) {
-            WizardStep.Basics -> renderBasicsStep()
-            WizardStep.Names -> renderNamesStep()
-            WizardStep.BirthDate -> renderBirthDateStep()
-            WizardStep.Contact -> renderContactStep()
-            WizardStep.LanguageLevel -> renderLanguageLevelStep()
-            is WizardStep.Question -> renderQuestion(
-                step.question,
-                step.questionIndex,
-                questionTranslations[step.questionIndex],
-            )
-        }
-    }
-
-    private fun buildSteps(includeDetails: Boolean): List<WizardStep> {
-        return buildList {
-            add(WizardStep.Basics)
-            if (includeDetails) {
-                add(WizardStep.Names)
-                add(WizardStep.BirthDate)
-                add(WizardStep.Contact)
-                add(WizardStep.LanguageLevel)
-            }
-            questions.forEachIndexed { index, question ->
-                add(WizardStep.Question(question, index))
-            }
-        }
-    }
-
-    private fun applyProfileDefaultsForCourseContent() {
-        if (courseId.isNullOrBlank()) {
-            return
-        }
-        val profile = UserProfileDatabase.getInstance(requireContext()).getProfile() ?: return
-        var hasOptionalDetails = false
-        fun assignIfEmpty(current: String?, incoming: String?): String? {
-            return if (current.isNullOrBlank() && !incoming.isNullOrBlank()) {
-                hasOptionalDetails = true
-                incoming
-            } else {
-                current
-            }
-        }
-        respondent.gender = assignIfEmpty(respondent.gender, profile.gender)
-        respondent.firstName = assignIfEmpty(respondent.firstName, profile.firstName)
-        respondent.middleName = assignIfEmpty(respondent.middleName, profile.middleName)
-        respondent.lastName = assignIfEmpty(respondent.lastName, profile.lastName)
-        respondent.email = assignIfEmpty(respondent.email, profile.email)
-        respondent.phoneNumber = assignIfEmpty(respondent.phoneNumber, profile.phoneNumber)
-        respondent.language = assignIfEmpty(respondent.language, profile.language)
-        respondent.level = assignIfEmpty(respondent.level, profile.level)
-        if (respondent.birthDate.isNullOrBlank() && !profile.birthDate.isNullOrBlank()) {
-            respondent.birthDate = profile.birthDate
-            birthDateSelection = parseBirthDateIso(profile.birthDate)
-            hasOptionalDetails = true
-        }
-        val birthDateMillis = birthDateSelection
-        if (birthDateMillis != null) {
-            val calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC")).apply {
-                timeInMillis = birthDateMillis
-            }
-            val year = calendar.get(Calendar.YEAR)
-            if (respondent.birthYear == null) {
-                respondent.birthYear = year
-            }
-            if (respondent.age == null) {
-                val nowYear = Calendar.getInstance().get(Calendar.YEAR)
-                respondent.age = nowYear - year
-            }
-        }
-        if (hasOptionalDetails) {
-            respondent.additionalInfo = true
-        }
-    }
-
-    private fun renderQuestion(
-        question: SurveyQuestion,
-        index: Int,
-        translation: TranslatedQuestion?,
-    ): Pair<View, () -> Boolean> {
-        return when (question.type) {
-            "input" -> renderTextInputQuestion(index, false)
-            "textarea" -> renderTextInputQuestion(index, true)
-            "select" -> renderSingleChoiceQuestion(question, index, translation)
-            "selectMultiple" -> renderMultiChoiceQuestion(question, index, translation)
-            "ratingScale" -> renderRatingQuestion(question, index)
-            else -> renderTextInputQuestion(index, false)
-        }
-    }
-
-    private fun buildSubmissionPayload(): Pair<List<SubmissionAnswer>, Int> {
-        val answersPayload = questions.mapIndexed { index, question ->
-            val answer = answers[index] ?: throw IllegalStateException("Missing answer for question $index")
-            if (isExam) {
-                val correct = isAnswerCorrect(question, answer)
-                val grade = if (correct) question.marks ?: 1 else 0
-                val mistakes = if (correct) 0 else 1
-                val value = when (answer) {
-                    is SurveyAnswer.Text -> answer.value
-                    is SurveyAnswer.SingleChoice -> answer.choice?.toSubmissionValue()
-                    is SurveyAnswer.MultipleChoice -> answer.choices.map { it.toSubmissionValue() }
-                    is SurveyAnswer.Rating -> answer.score.toString()
-                }
-                SubmissionAnswer(
-                    value = value,
-                    mistakes = mistakes,
-                    passed = correct,
-                    grade = grade,
-                )
-            } else {
-                when (answer) {
-                    is SurveyAnswer.Text -> SubmissionAnswer(value = answer.value)
-                    is SurveyAnswer.SingleChoice -> SubmissionAnswer(
-                        value = answer.choice?.toSubmissionValue(),
-                    )
-                    is SurveyAnswer.MultipleChoice -> SubmissionAnswer(
-                        value = answer.choices.map { it.toSubmissionValue() },
-                    )
-                    is SurveyAnswer.Rating -> SubmissionAnswer(value = answer.score.toString())
-                }
-            }
-        }
-        val totalGrade = if (isExam) {
-            answersPayload.sumOf { it.grade }
-        } else {
-            0
-        }
-        return answersPayload to totalGrade
-    }
-
-    private fun submitSurvey() {
-        if (isSubmitting) return
-        val survey = document ?: return
-        val base = baseUrl?.takeIf { it.isNotBlank() }
-            ?: DashboardServerPreferences.getServerBaseUrl(requireContext().applicationContext)
-        if (base.isNullOrBlank()) {
-            showValidationMessage(R.string.dashboard_surveys_missing_server)
-            return
-        }
-
-        val (answersPayload, totalGrade) = try {
-            buildSubmissionPayload()
-        } catch (_: IllegalStateException) {
-            showValidationMessage(R.string.dashboard_survey_wizard_input_required)
-            return
-        }
-        if (isExam && !isExamPassing()) {
-            showValidationMessage(R.string.dashboard_exam_incorrect_answers)
-            return
-        }
-
-        if (baseUrlOverride != null && !includeUserContext) {
-            submitPublicSurvey(base, survey, answersPayload)
-            return
-        }
-
-        val profile = if (includeUserContext) {
-            UserProfileDatabase.getInstance(requireContext()).getProfile()
-        } else {
-            null
-        }
-        val username = (profile?.username ?: credentials?.username)
-            ?.takeIf { includeUserContext }
-        if (username.isNullOrBlank() && isExam) {
-            showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-            return
-        }
-        val fullName = listOfNotNull(profile?.firstName, profile?.middleName, profile?.lastName)
-            .joinToString(" ")
-            .trim()
-            .takeIf { it.isNotBlank() }
-            ?: username
-            ?: ANONYMOUS_RESPONDENT_NAME
-
-        val parentId = buildSubmissionParentId(survey)
-        if (parentId == null) {
-            showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-            return
-        }
-
-        viewLifecycleOwner.lifecycleScope.launch {
-            setSubmitting(true)
-            val existingSubmission = fetchExistingSubmissionOrNull(base, username, parentId, survey)
-            val submission = buildSurveySubmission(
-                SurveySubmissionParams(
-                    survey = survey,
-                    existingSubmission = existingSubmission,
-                    username = username,
-                    fullName = fullName,
-                    parentId = parentId,
-                    answersPayload = answersPayload,
-                    totalGrade = totalGrade,
-                    profile = profile
-                )
-            )
-            processSubmission(base, submission, survey, username)
-        }
-    }
-
-    private fun submitPublicSurvey(
-        base: String,
-        survey: SurveyDocument,
-        answersPayload: List<SubmissionAnswer>,
-    ) {
-        val publicTeamId = teamId ?: survey.teamId
-        val publicSurveyId = survey.id
-        if (publicTeamId.isNullOrBlank() || publicSurveyId.isNullOrBlank()) {
-            showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-            return
-        }
-        viewLifecycleOwner.lifecycleScope.launch {
-            setSubmitting(true)
-            val result = submissionRepository.submitPublicSurvey(
-                baseUrl = base,
-                teamId = publicTeamId,
-                surveyId = publicSurveyId,
-                answers = answersPayload.map { it.value },
-            )
-            setSubmitting(false)
-            if (result.isSuccess) {
-                Toast.makeText(
-                    requireContext(),
-                    getString(R.string.dashboard_survey_wizard_completed),
-                    Toast.LENGTH_SHORT,
-                ).show()
-                finishWithResult()
-            } else {
-                showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-            }
-        }
-    }
-
-    private suspend fun fetchExistingSubmissionOrNull(
-        base: String,
-        username: String?,
-        parentId: String,
-        survey: SurveyDocument
-    ): DashboardSurveySubmissionsRepository.SubmissionLookup? {
-        if (username.isNullOrBlank()) {
-            return null
-        }
-        return courseId?.let {
-            submissionRepository.fetchExistingSubmission(
-                base,
-                credentials,
-                sessionCookie,
-                parentId,
-                userId = if (isExam) null else "org.couchdb.user:$username",
-                userName = if (isExam) username else null,
-                parentRev = if (isExam) survey.rev else null,
-                type = if (isExam) "exam" else "survey",
-            ).getOrNull()
-        }
-    }
-
-    private fun buildSurveySubmission(
-        params: SurveySubmissionParams
-    ): SurveySubmission {
-        val survey = params.survey
-        val existingSubmission = params.existingSubmission
-        val username = params.username
-        val fullName = params.fullName
-        val parentId = params.parentId
-        val answersPayload = params.answersPayload
-        val totalGrade = params.totalGrade
-        val profile = params.profile
-
-        val rawProfile = parseProfileRawDocument(profile?.rawDocument)
-        val fallbackUserId = username?.let { "org.couchdb.user:$it" }
-        val resolvedUserId = rawProfile?.optString("_id").takeIf { !it.isNullOrBlank() } ?: fallbackUserId
-        val resolvedUserName = rawProfile?.optString("name").takeIf { !it.isNullOrBlank() }
-            ?: listOfNotNull(respondent.firstName, respondent.middleName, respondent.lastName)
-                .joinToString(" ")
-                .trim()
-                .takeIf { it.isNotBlank() }
-            ?: fullName
-        val resolvedPlanetCode = rawProfile?.optString("planetCode").takeIf { !it.isNullOrBlank() } ?: serverCode
-        val resolvedParentCode = rawProfile?.optString("parentCode").takeIf { !it.isNullOrBlank() } ?: parentCode
-        return SurveySubmission(
-            id = existingSubmission?.id,
-            rev = existingSubmission?.rev,
-            type = if (isExam) "exam" else "survey",
-            parentId = parentId,
-            parent = SubmissionParent(
-                id = survey.id,
-                rev = survey.rev,
-                name = survey.name,
-                type = if (isExam) "courses" else "surveys",
-                passingPercentage = survey.passingPercentage,
-                teamShareAllowed = if (isExam) null else false,
-                questions = survey.questions,
-                description = survey.description,
-            ),
-            user = DashboardSurveySubmissionsRepository.SubmissionUser(
-                id = resolvedUserId,
-                name = resolvedUserName,
-                planetCode = resolvedPlanetCode,
-                parentCode = resolvedParentCode,
-                firstName = respondent.firstName ?: rawProfile?.optString("firstName").takeIf { !it.isNullOrBlank() },
-                middleName = respondent.middleName ?: rawProfile?.optString("middleName").takeIf { !it.isNullOrBlank() },
-                lastName = respondent.lastName ?: rawProfile?.optString("lastName").takeIf { !it.isNullOrBlank() },
-                email = respondent.email ?: rawProfile?.optString("email").takeIf { !it.isNullOrBlank() },
-                language = respondent.language ?: rawProfile?.optString("language").takeIf { !it.isNullOrBlank() },
-                phoneNumber = respondent.phoneNumber ?: rawProfile?.optString("phoneNumber").takeIf { !it.isNullOrBlank() },
-                birthDate = respondent.birthDate ?: rawProfile?.optString("birthDate").takeIf { !it.isNullOrBlank() },
-                age = respondent.age ?: rawProfile.optIntOrNull("age"),
-                gender = respondent.gender ?: rawProfile?.optString("gender").takeIf { !it.isNullOrBlank() },
-                level = respondent.level ?: rawProfile?.optString("level").takeIf { !it.isNullOrBlank() },
-            ),
-            team = (teamId ?: survey.teamId)?.let { id ->
-                SubmissionTeam(
-                    id = id,
-                    name = teamName,
-                    type = "local",
-                )
-            },
-            answers = answersPayload,
-            grade = totalGrade,
-            status = if (isExam) resolveExamStatus(survey) else "complete",
-            startTime = startTimeMillis,
-            lastUpdateTime = System.currentTimeMillis(),
-            source = serverCode,
-            parentCode = parentCode,
-            deviceName = org.ole.planet.myplanet.lite.util.DeviceUtils.getDeviceName(),
-            customDeviceName = resolveCustomDeviceName(),
-        )
-    }
-
-    private fun parseProfileRawDocument(rawDocument: String?): JSONObject? {
-        if (rawDocument.isNullOrBlank()) return null
-        return runCatching { JSONObject(rawDocument) }.getOrNull()
-    }
-
-    private fun JSONObject?.optIntOrNull(key: String): Int? {
-        this ?: return null
-        if (!has(key)) return null
-        return when (val value = opt(key)) {
-            is Number -> value.toInt()
-            is String -> value.toIntOrNull()
-            else -> null
-        }
-    }
-
-    private suspend fun processSubmission(
-        base: String,
-        submission: SurveySubmission,
-        survey: SurveyDocument,
-        username: String?
-    ) {
-        val isOnline = NetworkUtils.isDeviceOnline(requireContext())
-        if (!isOnline) {
-            setSubmitting(false)
-            if (baseUrlOverride != null) {
-                showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-                return
-            }
-            queueSubmissionForOutbox(submission, survey)
-            return
-        }
-
-        val result = submissionRepository.submitSurvey(base, credentials, sessionCookie, submission)
-        setSubmitting(false)
-        if (result.isSuccess) {
-            DashboardSurveyStatusStore(
-                requireContext().applicationContext,
-                username,
-            ).markCompleted(survey.id)
-            Toast.makeText(
-                requireContext(),
-                getString(
-                    if (isExam) {
-                        R.string.dashboard_exam_completed
-                    } else {
-                        R.string.dashboard_survey_wizard_completed
-                    }
-                ),
-                Toast.LENGTH_SHORT
-            ).show()
-            finishWithResult()
-        } else {
-            if (baseUrlOverride != null) {
-                showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-                return
-            }
-            queueSubmissionForOutbox(submission, survey)
-        }
-    }
-
-    private fun queueSubmissionForOutbox(submission: SurveySubmission, survey: SurveyDocument) {
-        viewLifecycleOwner.lifecycleScope.launch {
-            val saved = localSurveyRepository.saveSubmission(
-                submission = submission,
-                surveyId = survey.id,
-                surveyName = survey.name,
-                teamId = teamId,
-                teamName = teamName,
-            )
-            if (saved) {
-                Toast.makeText(
-                    requireContext(),
-                    getString(R.string.dashboard_survey_submission_saved_offline),
-                    Toast.LENGTH_SHORT,
-                ).show()
-                finishWithResult()
-            } else {
-                showValidationMessage(R.string.dashboard_survey_wizard_submission_failed)
-            }
-        }
-    }
-
-    private suspend fun flushPendingSurveySubmissions() {
-        localSurveyRepository.flushPendingSurveyOutbox()
-    }
-
-    private fun buildSubmissionParentId(survey: SurveyDocument): String? {
-        val surveyId = survey.id?.takeIf { it.isNotBlank() } ?: return null
-        val course = courseId?.takeIf { it.isNotBlank() }
-        return if (course == null) {
-            surveyId
-        } else {
-            "$surveyId@$course"
-        }
-    }
-
-    private fun resolveExamStatus(survey: SurveyDocument): String {
-        val requiresGrading = survey.questions.orEmpty().any { question ->
-            question.type.equals("input", ignoreCase = true) ||
-                question.type.equals("textarea", ignoreCase = true)
-        }
-        return if (requiresGrading) "requires grading" else "complete"
-    }
-
-    private fun isExamPassing(): Boolean {
-        val survey = document ?: return true
-        val questions = survey.questions.orEmpty()
-        if (questions.isEmpty()) return true
-        val passingPercentage = survey.passingPercentage ?: DEFAULT_EXAM_PASSING_PERCENTAGE
-        val totalMarks = questions.sumOf { it.marks ?: 1 }
-        if (totalMarks == 0) return true
-        val earned = questions.mapIndexed { index, question ->
-            if (isAnswerCorrect(question, answers[index])) {
-                question.marks ?: 1
-            } else {
-                0
-            }
-        }.sum()
-        val percentage = (earned * 100.0 / totalMarks).roundToInt()
-        return percentage >= passingPercentage
-    }
-
-    private fun isAnswerCorrect(question: SurveyQuestion, answer: SurveyAnswer?): Boolean {
-        val correctChoice = question.correctChoice
-        val normalizedCorrect = normalizeCorrectChoice(correctChoice)
-        val choiceIds = question.choices.orEmpty()
-            .mapNotNull { it.id?.trim() }
-            .filter { it.isNotBlank() }
-            .toSet()
-        val choiceTextToId = question.choices.orEmpty()
-            .mapNotNull { choice ->
-                val text = choice.text?.trim().orEmpty()
-                val id = choice.id?.trim().orEmpty()
-                if (text.isNotBlank() && id.isNotBlank()) text to id else null
-            }
-            .toMap()
-        val correctIds = normalizedCorrect
-            .map { it.trim() }
-            .filter { it.isNotBlank() && choiceIds.contains(it) }
-        val correctTexts = if (correctIds.isEmpty()) {
-            normalizedCorrect.map { it.trim() }.filter { it.isNotBlank() }
-        } else {
-            emptyList()
-        }
-        return when (question.type) {
-            "input", "textarea" -> isTextInputCorrect(normalizedCorrect, answer)
-            "select" -> isSelectInputCorrect(correctIds, correctTexts, choiceTextToId, answer)
-            "selectMultiple" -> isSelectMultipleInputCorrect(correctIds, answer)
-            "ratingScale" -> isRatingScaleInputCorrect(normalizedCorrect, answer)
-            else -> false
-        }
-    }
-
-    private fun isTextInputCorrect(normalizedCorrect: List<String>, answer: SurveyAnswer?): Boolean {
-        val correctText = normalizedCorrect.firstOrNull()?.trim().orEmpty()
-        val response = (answer as? SurveyAnswer.Text)?.value?.trim().orEmpty()
-        return if (correctText.isBlank()) {
-            response.isNotBlank()
-        } else {
-            response.equals(correctText, ignoreCase = true)
-        }
-    }
-
-    private fun isSelectInputCorrect(
-        correctIds: List<String>,
-        correctTexts: List<String>,
-        choiceTextToId: Map<String, String>,
-        answer: SurveyAnswer?
-    ): Boolean {
-        val selectedChoice = (answer as? SurveyAnswer.SingleChoice)?.choice
-        return if (correctIds.isNotEmpty()) {
-            val selectedId = normalizeSelectedId(selectedChoice?.id)
-            val resolvedId = selectedId?.takeIf { it.isNotBlank() }
-                ?: selectedChoice?.text?.trim()?.let { choiceTextToId[it] }
-            resolvedId != null && correctIds.contains(resolvedId)
-        } else {
-            val selectedText = selectedChoice?.text?.trim()
-            selectedText != null && correctTexts.any { it.equals(selectedText, ignoreCase = true) }
-        }
-    }
-
-    private fun isSelectMultipleInputCorrect(correctIds: List<String>, answer: SurveyAnswer?): Boolean {
-        val selectedIds = (answer as? SurveyAnswer.MultipleChoice)?.choices
-            ?.filter { !it.isOther }
-            ?.mapNotNull { normalizeSelectedId(it.id) }
-            ?.filter { it.isNotBlank() }
-            .orEmpty()
-        return selectedIds.size == correctIds.size && selectedIds.toSet() == correctIds.toSet()
-    }
-
-    private fun isRatingScaleInputCorrect(normalizedCorrect: List<String>, answer: SurveyAnswer?): Boolean {
-        val correctValue = normalizedCorrect.firstOrNull()?.trim().orEmpty()
-        val selectedScore = (answer as? SurveyAnswer.Rating)?.score?.toString()
-        return correctValue.isNotBlank() && selectedScore == correctValue
-    }
-
-    private fun normalizeCorrectChoice(correctChoice: Any?): List<String> {
-        return when (correctChoice) {
-            is String -> listOf(correctChoice)
-            is List<*> -> correctChoice.mapNotNull { item ->
-                when (item) {
-                    is Map<*, *> -> item["id"]?.toString()
-                        ?: item["_id"]?.toString()
-                        ?: item["text"]?.toString()
-                    else -> item?.toString()
-                }
-            }
-            null -> emptyList()
-            else -> listOf(correctChoice.toString())
-        }
-    }
-
-    private fun normalizeSelectedId(rawId: String?): String? {
-        val trimmed = rawId?.trim().orEmpty()
-        if (trimmed.isBlank()) return null
-        return trimmed.substringBefore("/").trim().takeIf { it.isNotBlank() }
-    }
-
-    private fun resolveCustomDeviceName(): String? {
-        val prefs = SecurePreferencesProvider.getServerPreferences(requireContext().applicationContext)
-        return prefs.getString(KEY_DEVICE_CUSTOM_DEVICE_NAME, null)?.trim()?.takeIf { it.isNotBlank() }
-    }
-
-    private fun finishWithResult() {
-        if (isExam) {
-            requireActivity().setResult(Activity.RESULT_OK)
-        }
-        requireActivity().finish()
-    }
-
-
-
-    private fun setSubmitting(submitting: Boolean) {
-        isSubmitting = submitting
-        updateNavigationEnabled()
-    }
-
-    private fun createGenderGroup(context: android.content.Context): Pair<TextView, RadioGroup> {
-        val genderLabel = TextView(context).apply {
-            text = getString(R.string.dashboard_survey_wizard_gender_label)
-        }
-        val genderGroup = RadioGroup(context).apply {
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val maleButton = RadioButton(context).apply {
-            text = getString(R.string.signup_gender_option_male)
-            tag = GENDER_MALE
-            id = View.generateViewId()
-        }
-        val femaleButton = RadioButton(context).apply {
-            text = getString(R.string.signup_gender_option_female)
-            tag = GENDER_FEMALE
-            id = View.generateViewId()
-        }
-        genderGroup.addView(maleButton)
-        genderGroup.addView(femaleButton)
-        when (respondent.gender) {
-            GENDER_MALE -> maleButton.isChecked = true
-            GENDER_FEMALE -> femaleButton.isChecked = true
-        }
-        return genderLabel to genderGroup
-    }
-
-    private fun createBirthYearLayout(context: android.content.Context): Pair<TextInputLayout, TextInputEditText> {
-        val birthYearLayout = TextInputLayout(context).apply {
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        birthYearLayout.hint = getString(R.string.dashboard_survey_wizard_birth_year_label)
-        val birthYearInput = TextInputEditText(context).apply {
-            inputType = InputType.TYPE_CLASS_NUMBER
-            setText(respondent.birthYear?.toString().orEmpty())
-            setOnFocusChangeListener { view, hasFocus ->
-                if (hasFocus) {
-                    scrollFocusedFieldIntoView(view)
-                } else {
-                    updateRespondentBirthYear(text?.toString())
-                }
-            }
-            setOnClickListener { scrollFocusedFieldIntoView(this) }
-            imeOptions = EditorInfo.IME_ACTION_DONE
-            setOnEditorActionListener { _, actionId, event ->
-                val isDoneAction = actionId == EditorInfo.IME_ACTION_DONE ||
-                    event?.keyCode == KeyEvent.KEYCODE_ENTER
-                if (isDoneAction) {
-                    updateRespondentBirthYear(text?.toString())
-                }
-                false
-            }
-        }
-        birthYearLayout.addView(birthYearInput)
-        return birthYearLayout to birthYearInput
-    }
-
-    private fun scrollFocusedFieldIntoView(view: View) {
-        questionScrollView.post {
-            questionScrollView.postDelayed({
-                val scrollChild = questionScrollView.getChildAt(0) as? ViewGroup ?: return@postDelayed
-                val targetRect = Rect().also { view.getDrawingRect(it) }
-                scrollChild.offsetDescendantRectToMyCoords(view, targetRect)
-                targetRect.bottom += questionScrollView.height / 3
-                questionScrollView.requestChildRectangleOnScreen(scrollChild, targetRect, true)
-            }, 180)
-        }
-    }
-
-    private fun updateRespondentBirthYear(value: String?) {
-        val year = value?.trim().orEmpty().toIntOrNull()
-        respondent.birthYear = year
-        respondent.age = year?.let { Calendar.getInstance().get(Calendar.YEAR) - it }
-    }
-
-    private fun createAdditionalCheckBox(context: android.content.Context): CheckBox {
-        return CheckBox(context).apply {
-            text = getString(R.string.dashboard_survey_wizard_additional_info_label)
-            isChecked = respondent.additionalInfo
-        }
-    }
-
-    private fun renderBasicsStep(): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val (genderLabel, genderGroup) = createGenderGroup(context)
-        val (birthYearLayout, birthYearInput) = createBirthYearLayout(context)
-        val additionalCheckBox = createAdditionalCheckBox(context)
-
-        container.addView(genderLabel)
-        container.addView(genderGroup)
-        container.addView(birthYearLayout)
-        container.addView(additionalCheckBox)
-
-        val collector = {
-            val checkedId = genderGroup.checkedRadioButtonId
-            val yearText = birthYearInput.text?.toString()?.trim().orEmpty()
-            val year = yearText.toIntOrNull()
-            if (yearText.isNotEmpty() && year == null) {
-                showValidationMessage(R.string.dashboard_survey_wizard_birth_year_required)
-                false
-            } else {
-                val nowYear = Calendar.getInstance().get(Calendar.YEAR)
-                respondent.gender = genderGroup.findViewById<RadioButton>(checkedId)?.tag as? String
-                respondent.birthYear = year
-                respondent.age = year?.let { nowYear - it }
-                respondent.additionalInfo = additionalCheckBox.isChecked
-                if (includeOptionalDetails != respondent.additionalInfo) {
-                    includeOptionalDetails = respondent.additionalInfo
-                    steps = buildSteps(includeOptionalDetails)
-                }
-                true
-            }
-        }
-        return container to collector
-    }
-
-    private fun renderNamesStep(): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val firstLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_first_name_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val firstInput = TextInputEditText(context)
-        firstInput.setText(respondent.firstName.orEmpty())
-        firstLayout.addView(firstInput)
-
-        val middleLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_middle_name_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val middleInput = TextInputEditText(context)
-        middleInput.setText(respondent.middleName.orEmpty())
-        middleLayout.addView(middleInput)
-
-        val lastLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_last_name_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val lastInput = TextInputEditText(context)
-        lastInput.setText(respondent.lastName.orEmpty())
-        lastLayout.addView(lastInput)
-
-        container.addView(firstLayout)
-        container.addView(middleLayout)
-        container.addView(lastLayout)
-
-        val collector = {
-            respondent.firstName = firstInput.text?.toString()?.trim().orEmpty().takeIf { it.isNotBlank() }
-            respondent.middleName = middleInput.text?.toString()?.trim().orEmpty().takeIf { it.isNotBlank() }
-            respondent.lastName = lastInput.text?.toString()?.trim().orEmpty().takeIf { it.isNotBlank() }
-            true
-        }
-        return container to collector
-    }
-
-    private fun renderBirthDateStep(): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val birthDateLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_birth_date_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val birthDateInput = TextInputEditText(context)
-        birthDateInput.apply {
-            inputType = InputType.TYPE_CLASS_DATETIME or InputType.TYPE_DATETIME_VARIATION_DATE
-            keyListener = null
-            isFocusable = false
-            isClickable = true
-            setText(respondent.birthDate?.let { formatBirthDateDisplay(it) }
-                ?: birthDateSelection?.let { formatBirthDateIso(it) }.orEmpty())
-            setOnClickListener { showBirthDatePicker(this) }
-            setOnFocusChangeListener { _, hasFocus ->
-                if (hasFocus) {
-                    showBirthDatePicker(this)
-                }
-            }
-        }
-        birthDateLayout.addView(birthDateInput)
-        container.addView(birthDateLayout)
-
-        val collector = {
-            respondent.birthDate = birthDateSelection?.let { formatBirthDateIso(it) }
-            true
-        }
-        return container to collector
-    }
-
-    private fun showBirthDatePicker(input: TextInputEditText) {
-        if (childFragmentManager.findFragmentByTag(BIRTH_DATE_PICKER_TAG) != null) {
-            return
-        }
-
-        val picker = MaterialDatePicker.Builder.datePicker()
-            .setTitleText(getString(R.string.signup_birth_date_picker_title))
-            .apply {
-                birthDateSelection?.let { setSelection(it) }
-            }
-            .build()
-
-        picker.addOnPositiveButtonClickListener { selection ->
-            birthDateSelection = selection
-            input.setText(formatBirthDateIso(selection))
-        }
-
-        picker.addOnDismissListener {
-            input.clearFocus()
-        }
-
-        picker.show(childFragmentManager, BIRTH_DATE_PICKER_TAG)
-    }
-
-    private fun parseBirthDateIso(value: String?): Long? {
-        if (value.isNullOrBlank()) {
-            return null
-        }
-        val formatter = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).apply {
-            timeZone = TimeZone.getTimeZone("UTC")
-        }
-        return try {
-            formatter.parse(value)?.time
-        } catch (_: ParseException) {
-            null
-        }
-    }
-
-    private fun formatBirthDateIso(selection: Long): String {
-        val formatter = SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).apply {
-            timeZone = TimeZone.getTimeZone("UTC")
-        }
-        return formatter.format(Date(selection))
-    }
-
-    private fun formatBirthDateDisplay(value: String): String {
-        val parsed = parseBirthDateIso(value)
-        return parsed?.let { formatBirthDateIso(it) } ?: value
-    }
-
-    private fun renderContactStep(): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val emailLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_email_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val emailInput = TextInputEditText(context).apply {
-            inputType = InputType.TYPE_TEXT_VARIATION_EMAIL_ADDRESS
-            setText(respondent.email.orEmpty())
-        }
-        emailLayout.addView(emailInput)
-
-        val phoneLayout = TextInputLayout(context).apply {
-            hint = getString(R.string.dashboard_survey_wizard_phone_label)
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val phoneInput = TextInputEditText(context).apply {
-            inputType = InputType.TYPE_CLASS_PHONE
-            setText(respondent.phoneNumber.orEmpty())
-        }
-        phoneLayout.addView(phoneInput)
-
-        container.addView(emailLayout)
-        container.addView(phoneLayout)
-
-        val collector = {
-            respondent.email = emailInput.text?.toString()?.trim().orEmpty().takeIf { it.isNotBlank() }
-            respondent.phoneNumber = phoneInput.text?.toString()?.trim().orEmpty().takeIf { it.isNotBlank() }
-            true
-        }
-        return container to collector
-    }
-
-    private fun createDropdownLayout(context: android.content.Context, hintText: String): Pair<TextInputLayout, AutoCompleteTextView> {
-        val layout = TextInputLayout(context).apply {
-            hint = hintText
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val input = AutoCompleteTextView(context).apply {
-            inputType = InputType.TYPE_NULL
-            keyListener = null
-            setOnClickListener { showDropDown() }
-            setOnFocusChangeListener { _, hasFocus -> if (hasFocus) showDropDown() }
-        }
-        layout.addView(input)
-        return layout to input
-    }
-
-    private fun levelArrayForLanguage(languageLabel: String?): Int {
-        val normalized = languageLabel?.trim()?.lowercase(Locale.ROOT)
-        return when (normalized) {
-            getString(R.string.language_name_spanish).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_es
-            getString(R.string.language_name_french).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_fr
-            getString(R.string.language_name_portuguese).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_pt
-            getString(R.string.language_name_arabic).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_ar
-            getString(R.string.language_name_somali).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_so
-            getString(R.string.language_name_nepali).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_ne
-            getString(R.string.language_name_hindi).lowercase(Locale.ROOT) -> R.array.signup_level_options_language_hi
-            else -> R.array.signup_level_options_language_en
-        }
-    }
-
-    private fun renderLanguageLevelStep(): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val (languageLayout, languageInput) = createDropdownLayout(
-            context,
-            getString(R.string.dashboard_survey_wizard_language_label)
-        )
-        val (levelLayout, levelInput) = createDropdownLayout(
-            context,
-            getString(R.string.dashboard_survey_wizard_level_label)
-        )
-
-        val languages = resources.getStringArray(R.array.signup_language_options).toList()
-        val languageAdapter = ArrayAdapter(context, android.R.layout.simple_spinner_dropdown_item, languages)
-        languageInput.setAdapter(languageAdapter)
-        respondent.language?.let { languageInput.setText(it, false) }
-
-        var currentLevelOptions: List<String> = emptyList()
-
-        fun updateLevelOptions(languageLabel: String?) {
-            val arrayRes = levelArrayForLanguage(languageLabel)
-            val options = resources.getStringArray(arrayRes).toList()
-            currentLevelOptions = options
-            val adapter = ArrayAdapter(context, android.R.layout.simple_spinner_dropdown_item, options)
-            levelInput.setAdapter(adapter)
-            val localizedLevel = LearningLevelTranslator.toLocalized(context, respondent.level, arrayRes)
-            if (!localizedLevel.isNullOrBlank() && options.contains(localizedLevel)) {
-                levelInput.setText(localizedLevel, false)
-            } else {
-                levelInput.setText("", false)
-            }
-        }
-
-        languageInput.setOnItemClickListener { _, _, position, _ ->
-            val selected = languageAdapter.getItem(position)
-            updateLevelOptions(selected)
-        }
-
-        updateLevelOptions(respondent.language)
-
-        container.addView(languageLayout)
-        container.addView(levelLayout)
-
-        val collector = {
-            val selectedLanguage = languageInput.text?.toString()?.trim().orEmpty()
-            respondent.language = selectedLanguage.takeIf { languages.contains(it) }
-
-            val levelText = levelInput.text?.toString()?.trim().orEmpty()
-            respondent.level = if (currentLevelOptions.contains(levelText)) {
-                LearningLevelTranslator.toEnglish(context, levelText)
-            } else {
-                null
-            }
-            true
-        }
-        return container to collector
-    }
-
-    private fun renderTextInputQuestion(index: Int, multiline: Boolean): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val layout = TextInputLayout(context)
-        layout.layoutParams = LinearLayout.LayoutParams(
-            ViewGroup.LayoutParams.MATCH_PARENT,
-            ViewGroup.LayoutParams.WRAP_CONTENT,
-        )
-        val editText = TextInputEditText(context)
-        editText.layoutParams = LinearLayout.LayoutParams(
-            ViewGroup.LayoutParams.MATCH_PARENT,
-            ViewGroup.LayoutParams.WRAP_CONTENT,
-        )
-        editText.isSingleLine = !multiline
-        if (multiline) {
-            editText.minLines = 3
-            editText.inputType = InputType.TYPE_CLASS_TEXT or InputType.TYPE_TEXT_FLAG_MULTI_LINE
-        } else {
-            editText.inputType = InputType.TYPE_CLASS_TEXT
-        }
-        val savedText = (answers[index] as? SurveyAnswer.Text)?.value.orEmpty()
-        if (savedText.isNotEmpty()) {
-            editText.setText(savedText)
-            editText.setSelection(savedText.length)
-        }
-        layout.addView(editText)
-        val collector = {
-            val response = editText.text?.toString()?.trim().orEmpty()
-            if (response.isBlank()) {
-                showValidationMessage(R.string.dashboard_survey_wizard_input_required)
-                false
-            } else {
-                answers[index] = SurveyAnswer.Text(response)
-                true
-            }
-        }
-        return layout to collector
-    }
-
-    private fun renderSingleChoiceQuestion(
-        question: SurveyQuestion,
-        index: Int,
-        translation: TranslatedQuestion?,
-    ): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val radioGroup = RadioGroup(context).apply {
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-
-        val otherInputLayout = buildSingleChoiceOptions(
-            context = context,
-            question = question,
-            translation = translation,
-            radioGroup = radioGroup,
-            container = container,
-        )
-        container.addView(radioGroup, 0)
-
-        val savedSelection = (answers[index] as? SurveyAnswer.SingleChoice)?.choice
-        restoreSingleChoiceSelection(radioGroup, otherInputLayout, savedSelection)
-
-        val collector = {
-            collectSingleChoiceAnswer(radioGroup, otherInputLayout, question, index)
-        }
-        return container to collector
-    }
-
-    private fun buildSingleChoiceOptions(
-        context: android.content.Context,
-        question: SurveyQuestion,
-        translation: TranslatedQuestion?,
-        radioGroup: RadioGroup,
-        container: LinearLayout,
-    ): TextInputLayout? {
-        val choices = question.choices.orEmpty()
-        choices.forEachIndexed { index, choice ->
-            val button = RadioButton(context)
-            val translatedLabel = translation?.choices?.getOrNull(index)
-            button.text = translatedLabel?.takeIf { it.isNotBlank() } ?: choice.text.orEmpty()
-            button.tag = choice
-            radioGroup.addView(button)
-        }
-        var otherInputLayout: TextInputLayout? = null
-        if (question.hasOtherOption) {
-            val otherButton = RadioButton(context)
-            otherButton.text = getString(R.string.dashboard_survey_wizard_other_option)
-            otherButton.tag = OTHER_CHOICE_TAG
-            radioGroup.addView(otherButton)
-            otherInputLayout = buildOtherInputField(context)
-            otherInputLayout.isVisible = false
-            container.addView(otherInputLayout)
-            radioGroup.setOnCheckedChangeListener { _, checkedId ->
-                val selectedButton = radioGroup.findViewById<RadioButton>(checkedId)
-                val isOther = selectedButton?.tag == OTHER_CHOICE_TAG
-                otherInputLayout.isVisible = isOther
-            }
-        }
-        return otherInputLayout
-    }
-
-    private fun restoreSingleChoiceSelection(
-        radioGroup: RadioGroup,
-        otherInputLayout: TextInputLayout?,
-        savedSelection: SelectedOption?,
-    ) {
-        if (savedSelection != null) {
-            val matchedButton = (0 until radioGroup.childCount)
-                .mapNotNull { childIndex -> radioGroup.getChildAt(childIndex) as? RadioButton }
-                .firstOrNull { button ->
-                    val taggedChoice = button.tag as? SurveyChoice
-                    if (savedSelection.isOther) {
-                        button.tag == OTHER_CHOICE_TAG
-                    } else {
-                        taggedChoice?.id == savedSelection.id ||
-                            button.text?.toString() == savedSelection.text
-                    }
-                }
-            matchedButton?.isChecked = true
-            if (savedSelection.isOther) {
-                otherInputLayout?.isVisible = true
-                otherInputLayout?.editText?.setText(savedSelection.text)
-                otherInputLayout?.editText?.setSelection(savedSelection.text.length)
-            }
-        }
-    }
-
-    private fun collectSingleChoiceAnswer(
-        radioGroup: RadioGroup,
-        otherInputLayout: TextInputLayout?,
-        question: SurveyQuestion,
-        index: Int,
-    ): Boolean {
-        val selectedId = radioGroup.checkedRadioButtonId
-        if (selectedId == -1) {
-            showValidationMessage(R.string.dashboard_survey_wizard_choice_required)
-            return false
-        }
-        val selectedButton = radioGroup.findViewById<RadioButton>(selectedId)
-        val isOther = selectedButton.tag == OTHER_CHOICE_TAG
-        val otherValue = otherInputLayout?.editText?.text?.toString()?.trim().orEmpty()
-        if (isOther && otherValue.isBlank()) {
-            showValidationMessage(R.string.dashboard_survey_wizard_input_required)
-            return false
-        }
-        val choice = if (isOther) {
-            SelectedOption(
-                id = GENDER_OTHER,
-                text = otherValue,
-                isOther = true,
-            )
-        } else {
-            val originalChoice = selectedButton.tag as? SurveyChoice
-            val label = originalChoice?.text?.takeIf { it.isNotBlank() }
-                ?: selectedButton.text?.toString()?.takeIf { it.isNotBlank() }
-                ?: originalChoice?.id
-                ?: ""
-            SelectedOption(
-                id = originalChoice?.id ?: label,
-                text = label,
-                isOther = false,
-            )
-        }
-        val answer = SurveyAnswer.SingleChoice(choice = choice)
-        if (isExam && !isAnswerCorrect(question, answer)) {
-            showValidationMessage(R.string.dashboard_exam_incorrect_answers)
-            return false
-        }
-        answers[index] = answer
-        return true
-    }
-
-    private fun renderMultiChoiceQuestion(
-        question: SurveyQuestion,
-        index: Int,
-        translation: TranslatedQuestion?,
-    ): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val container = LinearLayout(context).apply {
-            orientation = LinearLayout.VERTICAL
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-        val checkboxes = mutableListOf<android.widget.CheckBox>()
-        buildMultiChoiceCheckboxes(context, question, translation, container, checkboxes)
-
-        val (otherBox, otherInputLayout) = buildMultiChoiceOtherOption(context, question, container)
-
-        val savedSelections = (answers[index] as? SurveyAnswer.MultipleChoice)?.choices.orEmpty()
-        restoreMultiChoiceSelections(checkboxes, otherBox, otherInputLayout, savedSelections)
-
-        val collector = {
-            collectMultiChoiceAnswer(checkboxes, otherBox, otherInputLayout, question, index)
-        }
-        return container to collector
-    }
-
-    private fun buildMultiChoiceCheckboxes(
-        context: android.content.Context,
-        question: SurveyQuestion,
-        translation: TranslatedQuestion?,
-        container: LinearLayout,
-        checkboxes: MutableList<android.widget.CheckBox>,
-    ) {
-        question.choices.orEmpty().forEachIndexed { index, choice ->
-            val box = android.widget.CheckBox(context)
-            val translatedLabel = translation?.choices?.getOrNull(index)
-            box.text = translatedLabel?.takeIf { it.isNotBlank() } ?: choice.text.orEmpty()
-            box.tag = choice
-            container.addView(box)
-            checkboxes.add(box)
-        }
-    }
-
-    private fun buildMultiChoiceOtherOption(
-        context: android.content.Context,
-        question: SurveyQuestion,
-        container: LinearLayout,
-    ): Pair<android.widget.CheckBox?, TextInputLayout?> {
-        if (!question.hasOtherOption) return null to null
-        val otherBox = android.widget.CheckBox(context)
-        otherBox.text = getString(R.string.dashboard_survey_wizard_other_option)
-        otherBox.tag = OTHER_CHOICE_TAG
-        container.addView(otherBox)
-        val otherInputLayout = buildOtherInputField(context)
-        otherInputLayout.isVisible = false
-        container.addView(otherInputLayout)
-        otherBox.setOnCheckedChangeListener { _, isChecked ->
-            otherInputLayout.isVisible = isChecked
-        }
-        return otherBox to otherInputLayout
-    }
-
-    private fun restoreMultiChoiceSelections(
-        checkboxes: List<android.widget.CheckBox>,
-        otherBox: android.widget.CheckBox?,
-        otherInputLayout: TextInputLayout?,
-        savedSelections: List<SelectedOption>,
-    ) {
-        checkboxes.forEach { box ->
-            val choice = box.tag as? SurveyChoice
-            val isSelected = savedSelections.any { saved ->
-                saved.id == choice?.id || saved.text == box.text?.toString()
-            }
-            box.isChecked = isSelected
-        }
-        val savedOther = savedSelections.firstOrNull { it.isOther }
-        if (savedOther != null) {
-            otherBox?.isChecked = true
-            otherInputLayout?.isVisible = true
-            otherInputLayout?.editText?.setText(savedOther.text)
-            otherInputLayout?.editText?.setSelection(savedOther.text.length)
-        }
-    }
-
-    private fun collectMultiChoiceAnswer(
-        checkboxes: List<android.widget.CheckBox>,
-        otherBox: android.widget.CheckBox?,
-        otherInputLayout: TextInputLayout?,
-        question: SurveyQuestion,
-        index: Int,
-    ): Boolean {
-        val selectedChoices = checkboxes.filter { it.isChecked }
-            .map { checkbox ->
-                val choice = checkbox.tag as? SurveyChoice
-                val label = choice?.text?.takeIf { it.isNotBlank() }
-                    ?: checkbox.text?.toString()?.takeIf { it.isNotBlank() }
-                    ?: choice?.id
-                    ?: ""
-                SelectedOption(
-                    id = choice?.id ?: label,
-                    text = label,
-                    isOther = false,
-                )
-            }
-        val otherChecked = otherBox?.isChecked == true
-        val otherText = otherInputLayout?.editText?.text?.toString()?.trim().orEmpty()
-        if (selectedChoices.isEmpty() && !otherChecked) {
-            showValidationMessage(R.string.dashboard_survey_wizard_choice_required)
-            return false
-        } else if (otherChecked && otherText.isBlank()) {
-            showValidationMessage(R.string.dashboard_survey_wizard_input_required)
-            return false
-        }
-        val combined = mutableListOf<SelectedOption>()
-        combined.addAll(selectedChoices)
-        if (otherChecked && otherText.isNotBlank()) {
-            combined.add(
-                SelectedOption(
-                    id = GENDER_OTHER,
-                    text = otherText,
-                    isOther = true,
-                ),
-            )
-        }
-        val answer = SurveyAnswer.MultipleChoice(choices = combined)
-        if (isExam && !isAnswerCorrect(question, answer)) {
-            showValidationMessage(R.string.dashboard_exam_incorrect_answers)
-            return false
-        }
-        answers[index] = answer
-        return true
-    }
-
-    private fun renderRatingQuestion(question: SurveyQuestion, index: Int): Pair<View, () -> Boolean> {
-        val context = requireContext()
-        val scaleMax = question.scaleMax?.takeIf { it > 0 } ?: DEFAULT_RATING_SCALE_MAX
-        val gridLayout = createRatingGridLayout(context, scaleMax)
-
-        val styles = getRatingStyles(context)
-        val buttons = mutableListOf<MaterialButton>()
-        var selectedValue: Int? = (answers[index] as? SurveyAnswer.Rating)?.score
-
-        fun applySelection() {
-            buttons.forEachIndexed { idx, button ->
-                updateRatingButtonSelection(button, idx + 1, selectedValue == idx + 1, scaleMax, styles)
-            }
-        }
-
-        (1..scaleMax).forEach { value ->
-            val button = createRatingButton(context, value, styles.horizontalMargin, styles.bottomMargin) { selected ->
-                selectedValue = selected
-                applySelection()
-            }
-            buttons.add(button)
-            gridLayout.addView(button)
-        }
-
-        applySelection()
-
-        return gridLayout to createRatingCollector(index) { selectedValue }
-    }
-
-    private fun createRatingGridLayout(context: android.content.Context, scaleMax: Int): GridLayout {
-        val columnCount = ratingScaleColumnCount(scaleMax)
-        return GridLayout(context).apply {
-            rowCount = ((scaleMax + columnCount - 1) / columnCount).coerceAtLeast(1)
-            this.columnCount = columnCount
-            layoutParams = LinearLayout.LayoutParams(
-                ViewGroup.LayoutParams.MATCH_PARENT,
-                ViewGroup.LayoutParams.WRAP_CONTENT,
-            )
-        }
-    }
-
-    private data class RatingStyles(
-        val basePalette: List<Int>,
-        val selectedColor: Int,
-        val selectedTextColor: Int,
-        val defaultTextColor: Int,
-        val horizontalMargin: Int,
-        val bottomMargin: Int
-    )
-
-    private fun getRatingStyles(context: android.content.Context): RatingStyles {
-        return RatingStyles(
-            basePalette = listOf(
-                "#F3C4E3".toColorInt(),
-                "#F6CEDE".toColorInt(),
-                "#F8D7DA".toColorInt(),
-                "#FAE0D6".toColorInt(),
-                "#FDEAD1".toColorInt(),
-                "#FFF3CD".toColorInt(),
-                "#F1F1D1".toColorInt(),
-                "#E2EFD6".toColorInt(),
-                "#D4EDDA".toColorInt(),
-            ),
-            selectedColor = ContextCompat.getColor(context, R.color.survey_rating_selected_background),
-            selectedTextColor = ContextCompat.getColor(context, R.color.survey_rating_selected_text),
-            defaultTextColor = ContextCompat.getColor(context, R.color.survey_rating_default_text),
-            horizontalMargin = resources.getDimensionPixelSize(R.dimen.padding_small),
-            bottomMargin = resources.getDimensionPixelSize(R.dimen.padding_small)
-        )
-    }
-
-    private fun updateRatingButtonSelection(
-        button: MaterialButton,
-        value: Int,
-        isSelected: Boolean,
-        scaleMax: Int,
-        styles: RatingStyles
-    ) {
-        val backgroundColor = if (isSelected) {
-            styles.selectedColor
-        } else {
-            ratingBackgroundColor(value, scaleMax, styles.basePalette)
-        }
-        button.backgroundTintList = ColorStateList.valueOf(backgroundColor)
-        button.setTextColor(if (isSelected) styles.selectedTextColor else styles.defaultTextColor)
-        button.strokeWidth = 0
-    }
-
-    private fun createRatingCollector(index: Int, getSelectedValue: () -> Int?): () -> Boolean {
-        return {
-            val score = getSelectedValue()
-            if (score == null) {
-                showValidationMessage(R.string.dashboard_survey_wizard_rating_required)
-                false
-            } else {
-                answers[index] = SurveyAnswer.Rating(score)
-                true
-            }
-        }
-    }
-
-    private fun ratingScaleColumnCount(scaleMax: Int): Int {
-        return when {
-            scaleMax <= 5 -> scaleMax.coerceAtLeast(1)
-            scaleMax <= DEFAULT_RATING_SCALE_MAX -> 3
-            else -> kotlin.math.ceil(kotlin.math.sqrt(scaleMax.toDouble())).toInt().coerceAtLeast(1)
-        }
-    }
-
-    private fun ratingBackgroundColor(value: Int, scaleMax: Int, palette: List<Int>): Int {
-        if (palette.isEmpty()) return Color.TRANSPARENT
-        if (scaleMax <= 1 || palette.size == 1) return palette.first()
-        val normalizedPosition = (value - 1).coerceAtLeast(0).toFloat() / (scaleMax - 1)
-        val palettePosition = normalizedPosition * (palette.size - 1)
-        val lowerIndex = palettePosition.toInt().coerceIn(0, palette.lastIndex)
-        val upperIndex = minOf(lowerIndex + 1, palette.lastIndex)
-        val fraction = palettePosition - lowerIndex
-        return blendColors(palette[lowerIndex], palette[upperIndex], fraction)
-    }
-
-    private fun blendColors(startColor: Int, endColor: Int, fraction: Float): Int {
-        val red = (Color.red(startColor) + (Color.red(endColor) - Color.red(startColor)) * fraction).roundToInt()
-        val green = (Color.green(startColor) + (Color.green(endColor) - Color.green(startColor)) * fraction).roundToInt()
-        val blue = (Color.blue(startColor) + (Color.blue(endColor) - Color.blue(startColor)) * fraction).roundToInt()
-        return Color.rgb(red, green, blue)
-    }
-
-    private fun createRatingButton(
-        context: android.content.Context,
-        value: Int,
-        horizontalMargin: Int,
-        bottomMargin: Int,
-        onClick: (Int) -> Unit
-    ): MaterialButton {
-        return MaterialButton(context).apply {
-            text = value.toString()
-            isAllCaps = false
-            textSize = 18f
-            cornerRadius = resources.getDimensionPixelSize(R.dimen.padding_small)
-            insetTop = 0
-            insetBottom = 0
-            layoutParams = GridLayout.LayoutParams().apply {
-                width = 0
-                height = ViewGroup.LayoutParams.WRAP_CONTENT
-                columnSpec = GridLayout.spec(GridLayout.UNDEFINED, 1f)
-                setMargins(horizontalMargin, 0, horizontalMargin, bottomMargin)
-            }
-            setOnClickListener {
-                onClick(value)
-            }
-        }
-    }
-
-    private fun buildOtherInputField(context: android.content.Context): TextInputLayout {
-        val layout = TextInputLayout(context)
-        layout.layoutParams = LinearLayout.LayoutParams(
-            ViewGroup.LayoutParams.MATCH_PARENT,
-            ViewGroup.LayoutParams.WRAP_CONTENT,
-        )
-        val editText = TextInputEditText(context)
-        editText.hint = getString(R.string.dashboard_survey_wizard_other_hint)
-        editText.layoutParams = LinearLayout.LayoutParams(
-            ViewGroup.LayoutParams.MATCH_PARENT,
-            ViewGroup.LayoutParams.WRAP_CONTENT,
-        )
-        layout.addView(editText)
-        return layout
-    }
-
-    private fun showValidationMessage(@StringRes messageRes: Int) {
-        Toast.makeText(requireContext(), getString(messageRes), Toast.LENGTH_SHORT).show()
     }
 
     companion object {
@@ -1879,12 +197,6 @@ class SurveyWizardFragment : Fragment(R.layout.fragment_survey_wizard) {
         private const val ARG_IS_EXAM = "arg_is_exam"
         private const val ARG_BASE_URL = "arg_base_url"
         private const val ARG_INCLUDE_USER_CONTEXT = "arg_include_user_context"
-        private const val OTHER_CHOICE_TAG = "other_choice"
-        private const val BIRTH_DATE_PICKER_TAG = "survey_birth_date_picker"
-        private const val DEFAULT_RATING_SCALE_MAX = 9
-        private const val DEFAULT_EXAM_PASSING_PERCENTAGE = 100
-        private const val KEY_DEVICE_CUSTOM_DEVICE_NAME = "device_custom_device_name"
-        private const val ANONYMOUS_RESPONDENT_NAME = "Anonymous"
 
         fun newInstance(
             document: SurveyDocument,
@@ -1909,80 +221,3 @@ class SurveyWizardFragment : Fragment(R.layout.fragment_survey_wizard) {
         }
     }
 }
-
-private sealed class WizardStep : java.io.Serializable {
-    object Basics : WizardStep() {
-        private fun readResolve(): Any = Basics
-    }
-
-    object Names : WizardStep() {
-        private fun readResolve(): Any = Names
-    }
-
-    object BirthDate : WizardStep() {
-        private fun readResolve(): Any = BirthDate
-    }
-
-    object Contact : WizardStep() {
-        private fun readResolve(): Any = Contact
-    }
-
-    object LanguageLevel : WizardStep() {
-        private fun readResolve(): Any = LanguageLevel
-    }
-
-    data class Question(val question: SurveyQuestion, val questionIndex: Int) : WizardStep()
-}
-
-private data class SurveyRespondent(
-    var gender: String? = null,
-    var birthYear: Int? = null,
-    var age: Int? = null,
-    var additionalInfo: Boolean = false,
-    var firstName: String? = null,
-    var middleName: String? = null,
-    var lastName: String? = null,
-    var birthDate: String? = null,
-    var email: String? = null,
-    var phoneNumber: String? = null,
-    var language: String? = null,
-    var level: String? = null,
-)
-
-private sealed class SurveyAnswer : java.io.Serializable {
-    data class Text(val value: String) : SurveyAnswer()
-    data class SingleChoice(val choice: SelectedOption?) : SurveyAnswer()
-    data class MultipleChoice(val choices: List<SelectedOption>) : SurveyAnswer()
-    data class Rating(val score: Int) : SurveyAnswer()
-}
-
-private data class SelectedOption(
-    val id: String?,
-    val text: String,
-    val isOther: Boolean = false,
-) : java.io.Serializable {
-    fun toSubmissionValue(): SubmissionOptionValue {
-        return SubmissionOptionValue(
-            id = id,
-            text = text,
-            isOther = isOther,
-        )
-    }
-}
-
-private data class SubmissionOptionValue(
-    @param:Json(name = "id") val id: String?,
-    @param:Json(name = "text") val text: String?,
-    @param:Json(name = "isOther") val isOther: Boolean = false,
-)
-
-private data class SurveySubmissionParams(
-    val survey: SurveyDocument,
-    val existingSubmission: DashboardSurveySubmissionsRepository.SubmissionLookup?,
-    val username: String?,
-    val fullName: String,
-    val parentId: String,
-    val answersPayload: List<SubmissionAnswer>,
-    val totalGrade: Int,
-    val profile: UserProfile?
-)
